Moved image pipeline functions to dataset module
