+++ few_shot/experiments/fashion/hyperparameter_search.py
@@ -6,8 +6,7 @@ import skopt
 import tensorflow as tf
 
 from few_shot.experiments.fashion import evaluate_fashion_few_shot, fashion_dfs
-from few_shot.experiments.fashion.data_augmentation import augmented_img_pipeline_fn
-
+from few_shot.dataset.image_pipeline import augmented_img_pipeline_fn
 
 MAX_SHOTS = 5
